converted page.tsx from client to server side
