+++ src/app/admin/components/sales-chart.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import { Card, CardContent, Typography } from '@mui/material';
 import {
