+++ src/app/admin/components/side-bar.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import {
   Box,
