+++ src/app/admin/components/user-stats-card.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import React from 'react';
 import { Card, CardContent, Typography, Box } from '@mui/material';
 import {
